requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
bleach==6.1.0
cachetools==5.3.2
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.2
cron-descriptor==1.4.0
distlib==0.3.8
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.15
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==20.1.0
filelock==3.13.1
flake8==6.1.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.4
mccabe==0.7.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pyproject-api==1.6.1
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tox==4.11.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
